gitignore

+++ src/entity/particle/particle.h
@@ -4,7 +4,7 @@
 
 class Particle : public Entity {
     public:
-        Particle(vec2 pos, vec2 dir, float size, float speed, float lifespan, ofColor col);
+        Particle(vec2 pos, vec2 dir, float size, float speed, unsigned short lifespan, ofColor col);
 
         void update(double deltaTime) override;
         void draw() override;
@@ -15,7 +15,7 @@ class Particle : public Entity {
         vec2 _direction;
 
         float _lifetime;
-        float _lifespan;
+        unsigned short _lifespan; // we keep this as an unsigned short so that we can pass -1 as an argument to mean practically infinite
 
         ofColor _color;
